requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.5
django-appconf==1.0.6
django-encrypted-model-fields==0.6.5
django-jet-reboot==1.3.10
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.9.0
